Improved FFT analysis
subtracted off the mean value of the signal. When you perform an FFT on
a signal that isn't centered about 0 you get a compononent of your FFT
at 0HZ which accounts for the shift of the data.

I also made a note about another signal at 2.18e-10 HZ, which
corresponds to the period over which I have data.